Unt3 projects
The code committed is for Hostos GD105 project unit 3, which covers three projects. The first project is using arcs to mimic the hours, minutes, and seconds of the clock. The clock is designed for AM and PM portions. The second project is to create the Moiré pattern by overlapping two sets of vertical and horizontal lines. The smooth, oscillating value using the sine function is generated, and it makes a wave-like pattern. The Ben-Day dots effect was also added. The third project used four groups of different colored spheres to check the collision effect of these spheres. The independent class of sphere is built, which has the constructor and also moves, displays, and checks the collision functions. The Bounce off the edges and collision effect were checked from the x, y, and z axes as well.